Improve to be able to place flags after arguments (#36)
* Improve to be able to place flags after arguments

Currently stern does not work if flags are placed after arguments. With
this commit, it improve to be able to place flags after arguments like
kubectl.

```sh
stern .* -n kube-system
```

For this change, github.com/spf13/cobra is used instead of
gopkg.in/urfave/cli.v1 and help message is changed accordingly. This
commit is not breaking change, compatible.

github.com/spf13/cobra is released under the Apache 2.0 License.

Signed-off-by: Kazuki Suda <kazuki.suda@gmail.com>

* Update README.md

* Update examples in README.md