Allow to use stern as a kubectl plugin
Note that with this change, the archive files will no longer be wrapped
in a directory. Since the krew plugin manifest generated by GoReleaser
does not support wrapping in a directory, and  the krew-index
auto-approval does not work if the directory name contains a version.

**Before**

```
$ tar -tf stern_1.20.1_darwin_amd64.tar.gz
stern_1.20.1_darwin_amd64/LICENSE
stern_1.20.1_darwin_amd64/stern
```

**After**

```
$ tar -tf stern_1.20.1-SNAPSHOT-5240615_darwin_amd64.tar.gz
LICENSE
stern
```