Add `--stdin` for `stdin` log parsing (#292)
* Add --stdin flag

* Move Log and TailOptions to separate file

When Log and TailOptions are used by other classes it is now more
obvious that they're not hard lined to Tail.

* Create FileTail for tailing log files

FileTail is heavily based on Tail but I saw no obvious way to abstract
the common parts of the classes to not duplicate them. FileTail is so
small it's hardly noticeable anyway.

* Create stdin tailing when using --stdin

* Document --stdin flag

* Remove start/stop printing from FileTail

Also moving file tailing to main thread as there are no parallel tails
with that config.